3_RhinoV2.0_Gripper_control/requirements.txt
# RhinoV2.0 夹爪控制系统依赖库

# GUI框架
PyQt5>=5.15.0

# 数据处理和可视化
numpy>=1.19.0
matplotlib>=3.3.0

# 串口通信
pyserial>=3.5

# USB设备通信
pyusb>=1.2.0
